+++ src/app/common/values/errors.text.ts
@@ -15,4 +15,6 @@ export class ErrorTexts {
     public static readonly ERROR_UNKNOWN_AUTHENTICATION = "Unknown authentication error";
     public static readonly ERROR_DATASET_OPERATION_FAILED = "Dataset operation failed";
     public static readonly ERROR_ACCOUNT_IS_NOT_WHITELISTED = "Your account is not whitelisted";
+    public static readonly ERROR_ACCESS_TOKEN = "Access token error";
+    public static readonly ERROR_ACCESS_TOKEN_EXPIRED = "Access token expired";
 }